RECIPES!!!
Changed Modifier HasModifier, if none returns true, instead of
previously checking against the cut type, which was wrong.
Changed RecipeComponent.Verify() now bails on any check failure, making
it faster to Verify quickly.
Recipes now have a name, they only add themselves to the static list of
recipes if their name is unique.
Recipes must be attached to a RecipeUnity and placed in the world in
order to properly exist for recipe checks. This is gross, but I don't
know a way around this, in Unity.