Lots of neat stuff
God!
In control of all the environmentals, God knows what day it is, has the
day, and is in control of the directional light
Day
The day class keeps track of current day progress as well as total
number of days passed
PlayerBehaviour
Fixed nutritionQueued and poisonQueued to decrement as they were
intended to
Lights
Most light prefabs have been given the DayBasedLight behaviour.
DayBasedLight
Turns lights on when it becomes night
Can be given a variance that makes it so the lights have a random delay
before turning on
UI
Added some UI to the Home scene
UIInterface
The interface between UI and everything else
Static
Contains method calls for setting values on specific UI elements.